Fix computation of track average in Automixer
Currently, the track average is computed by adding the channels. This works most of the time, because the signal in the channels are similar (most often they’re mono, really), which is probably why this bug wasn’t noticed until now. But there are cases in which the channels are out-of-phase, and adding them together will result in destructive interference. Still, the channels are producing sound, and shouldn’t be muted by the automixer.

The solution is to add the channels **after squaring them**. This gets rid of the negative values.

In other words: we add up the channels on the power domain, not the amplitude domain.

See the example project on this Pull Request for a pathological case in which one of the aumixed microphones is a stereo signal in which the left and right channels are perfectly out-of-phase. The current automixer mutes that track. The fixed version I’m proposing mixes the track correctly.

See also http://replaygain.hydrogenaud.io/proposal/rms_energy.html